Python script to use RIRIS with boostlets
The steps for extrapolation and the use of matices LAMBDA and E has not been implemented.
Extrapolation was used to avoid artifacts due to the use of Fourier in a non-periodic image. Columns are added to left and right. (Maybe the image should be padded in columns and rows).
Matrix LAMBDA can be substituted for an operation where we set columns to 0 if they are not the measured data.
Matrix E can also be substituted by removing functions in the dictionary that we do not want to use to project.